qcacmn: Do not populate meta_hdr on Tx completion path for mesh mode
Do not populate the struct meta_hdr_s into for buffers received on
Tx completion path when operating in mesh mode.

Also update the fields in tx_capture_hdr as per the new definition.

+++ dp/wifi3.0/dp_txrx_wds.c
@@ -1226,7 +1226,7 @@ dp_get_completion_indication_for_stack(struct dp_soc *soc,
 	ppdu_hdr->last_msdu = last_msdu;
 	if (qdf_unlikely(pdev->latency_capture_enable)) {
 		ppdu_hdr->tsf = ts->tsf;
-		ppdu_hdr->time_latency = time_latency;
+		ppdu_hdr->time_latency = (uint32_t)time_latency;
 	}
 
 	return QDF_STATUS_SUCCESS;
